separate domain types from service
